CI: Build all targets on changes
This builds the target in case anything changes in target specific code.
This will build all subtargets and activate all boards.

This also builds the testing kernel if it is available.

If there is no external toolchain the toolchain will be build in the
target build process.

All the .github/workflows/target-*.yml files are automatically
generated. Building all targets takes multiple hours even with 20
builders, this only triggers a target build in case something changed
for this specific target. An own workflow file is added for each target
which triggers on pull requests and also earlier when some changes are
pushed to a branch on github. The labeler is not working on branch pushes.

This should help detect build problems when new boards are added or when
some changes are done to a target.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>